renaming cli folder to scripts
