+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup, find_packages
 setup(
     name='drserpent',
     version='2.4',
-    packages=['cli','init','init_example'],
+    packages=['scripts','init','init_example'],
     py_modules=['expect','runner','collector','executor'],
     include_package_data=True,
     install_requires=[
@@ -18,7 +18,7 @@ setup(
     zip_safe=False,
     entry_points={
         'console_scripts': [
-            'serpent=cli.main:cli'
+            'serpent=scripts.main:cli'
         ],
     }
     )
